Fix some crash/deadlock conditions that appears when changes stream reading position (slider) or changes the stream
When the user changes the stream or moves inside the timeframe with the slider, the input stream is reinitialized and opened again for reading. Meanwhile, the decoder loop does his job on its own thread, reading from the stream and giving the read data to the decoder and audio buffer. There is a race condition here, the reinitialization clashes with the reads (the old stream is used for reading).