Fix interference between stopped stream events with starting stream events
When a stream is stopped from the UI, its decoder loop is also stopped. Meanwhile, on another thread, the new starting stream is decoding its header and starts playing, sending events for the UI thread. If the stopping stream is slower than the starting stream, then the stopping stream events are sent last and the UI reflects that the stream just stopped, not that the new stream started. That's why the stopping thread events must be disabled.